+++ sei-ibc-go/modules/core/03-connection/types/msgs.go
@@ -39,16 +39,6 @@ func NewMsgConnectionOpenInit(
 	}
 }
 
-// Route implements sdk.Msg
-func (msg MsgConnectionOpenInit) Route() string {
-	return host.RouterKey
-}
-
-// Type implements sdk.Msg
-func (msg MsgConnectionOpenInit) Type() string {
-	return "connection_open_init"
-}
-
 // ValidateBasic implements sdk.Msg.
 func (msg MsgConnectionOpenInit) ValidateBasic() error {
 	if err := host.ClientIdentifierValidator(msg.ClientId); err != nil {
@@ -69,12 +59,6 @@ func (msg MsgConnectionOpenInit) ValidateBasic() error {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "string could not be parsed as address: %v", err)
 	}
 	return msg.Counterparty.ValidateBasic()
-}
-
-// GetSignBytes implements sdk.Msg. The function will panic since it is used
-// for amino transaction verification which IBC does not support.
-func (msg MsgConnectionOpenInit) GetSignBytes() []byte {
-	panic("IBC messages do not support amino")
 }
 
 // GetSigners implements sdk.Msg
@@ -112,16 +96,6 @@ func NewMsgConnectionOpenTry(
 		ConsensusHeight:      consensusHeight,
 		Signer:               signer,
 	}
-}
-
-// Route implements sdk.Msg
-func (msg MsgConnectionOpenTry) Route() string {
-	return host.RouterKey
-}
-
-// Type implements sdk.Msg
-func (msg MsgConnectionOpenTry) Type() string {
-	return "connection_open_try"
 }
 
 // ValidateBasic implements sdk.Msg
@@ -184,12 +158,6 @@ func (msg MsgConnectionOpenTry) UnpackInterfaces(unpacker codectypes.AnyUnpacker
 	return unpacker.UnpackAny(msg.ClientState, new(exported.ClientState))
 }
 
-// GetSignBytes implements sdk.Msg. The function will panic since it is used
-// for amino transaction verification which IBC does not support.
-func (msg MsgConnectionOpenTry) GetSignBytes() []byte {
-	panic("IBC messages do not support amino")
-}
-
 // GetSigners implements sdk.Msg
 func (msg MsgConnectionOpenTry) GetSigners() []sdk.AccAddress {
 	accAddr, err := sdk.AccAddressFromBech32(msg.Signer)
@@ -226,16 +194,6 @@ func NewMsgConnectionOpenAck(
 // UnpackInterfaces implements UnpackInterfacesMessage.UnpackInterfaces
 func (msg MsgConnectionOpenAck) UnpackInterfaces(unpacker codectypes.AnyUnpacker) error {
 	return unpacker.UnpackAny(msg.ClientState, new(exported.ClientState))
-}
-
-// Route implements sdk.Msg
-func (msg MsgConnectionOpenAck) Route() string {
-	return host.RouterKey
-}
-
-// Type implements sdk.Msg
-func (msg MsgConnectionOpenAck) Type() string {
-	return "connection_open_ack"
 }
 
 // ValidateBasic implements sdk.Msg
@@ -281,12 +239,6 @@ func (msg MsgConnectionOpenAck) ValidateBasic() error {
 	return nil
 }
 
-// GetSignBytes implements sdk.Msg. The function will panic since it is used
-// for amino transaction verification which IBC does not support.
-func (msg MsgConnectionOpenAck) GetSignBytes() []byte {
-	panic("IBC messages do not support amino")
-}
-
 // GetSigners implements sdk.Msg
 func (msg MsgConnectionOpenAck) GetSigners() []sdk.AccAddress {
 	accAddr, err := sdk.AccAddressFromBech32(msg.Signer)
@@ -310,16 +262,6 @@ func NewMsgConnectionOpenConfirm(
 	}
 }
 
-// Route implements sdk.Msg
-func (msg MsgConnectionOpenConfirm) Route() string {
-	return host.RouterKey
-}
-
-// Type implements sdk.Msg
-func (msg MsgConnectionOpenConfirm) Type() string {
-	return "connection_open_confirm"
-}
-
 // ValidateBasic implements sdk.Msg
 func (msg MsgConnectionOpenConfirm) ValidateBasic() error {
 	if !IsValidConnectionID(msg.ConnectionId) {
@@ -336,12 +278,6 @@ func (msg MsgConnectionOpenConfirm) ValidateBasic() error {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "string could not be parsed as address: %v", err)
 	}
 	return nil
-}
-
-// GetSignBytes implements sdk.Msg. The function will panic since it is used
-// for amino transaction verification which IBC does not support.
-func (msg MsgConnectionOpenConfirm) GetSignBytes() []byte {
-	panic("IBC messages do not support amino")
 }
 
 // GetSigners implements sdk.Msg