+++ sei-ibc-go/modules/core/04-channel/types/events.go
@@ -14,11 +14,12 @@ const (
 	AttributeCounterpartyPortID    = "counterparty_port_id"
 	AttributeCounterpartyChannelID = "counterparty_channel_id"
 
-	EventTypeSendPacket        = "send_packet"
-	EventTypeRecvPacket        = "recv_packet"
-	EventTypeWriteAck          = "write_acknowledgement"
-	EventTypeAcknowledgePacket = "acknowledge_packet"
-	EventTypeTimeoutPacket     = "timeout_packet"
+	EventTypeSendPacket           = "send_packet"
+	EventTypeRecvPacket           = "recv_packet"
+	EventTypeWriteAck             = "write_acknowledgement"
+	EventTypeAcknowledgePacket    = "acknowledge_packet"
+	EventTypeTimeoutPacket        = "timeout_packet"
+	EventTypeTimeoutPacketOnClose = "timeout_on_close_packet"
 
 	// NOTE: DEPRECATED in favor of AttributeKeyDataHex
 	AttributeKeyData = "packet_data"
@@ -38,12 +39,12 @@ const (
 
 // IBC channel events vars
 var (
-	EventTypeChannelOpenInit     = MsgChannelOpenInit{}.Type()
-	EventTypeChannelOpenTry      = MsgChannelOpenTry{}.Type()
-	EventTypeChannelOpenAck      = MsgChannelOpenAck{}.Type()
-	EventTypeChannelOpenConfirm  = MsgChannelOpenConfirm{}.Type()
-	EventTypeChannelCloseInit    = MsgChannelCloseInit{}.Type()
-	EventTypeChannelCloseConfirm = MsgChannelCloseConfirm{}.Type()
+	EventTypeChannelOpenInit     = "channel_open_init"
+	EventTypeChannelOpenTry      = "channel_open_try"
+	EventTypeChannelOpenAck      = "channel_open_ack"
+	EventTypeChannelOpenConfirm  = "channel_open_confirm"
+	EventTypeChannelCloseInit    = "channel_close_init"
+	EventTypeChannelCloseConfirm = "channel_close_confirm"
 
 	AttributeValueCategory = fmt.Sprintf("%s_%s", host.ModuleName, SubModuleName)
 )