+++ sei-ibc-go/modules/core/04-channel/types/msgs.go
@@ -28,16 +28,6 @@ func NewMsgChannelOpenInit(
 	}
 }
 
-// Route implements sdk.Msg
-func (msg MsgChannelOpenInit) Route() string {
-	return host.RouterKey
-}
-
-// Type implements sdk.Msg
-func (msg MsgChannelOpenInit) Type() string {
-	return "channel_open_init"
-}
-
 // ValidateBasic implements sdk.Msg
 func (msg MsgChannelOpenInit) ValidateBasic() error {
 	if err := host.PortIdentifierValidator(msg.PortId); err != nil {
@@ -57,12 +47,6 @@ func (msg MsgChannelOpenInit) ValidateBasic() error {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "string could not be parsed as address: %v", err)
 	}
 	return msg.Channel.ValidateBasic()
-}
-
-// GetSignBytes implements sdk.Msg. The function will panic since it is used
-// for amino transaction verification which IBC does not support.
-func (msg MsgChannelOpenInit) GetSignBytes() []byte {
-	panic("IBC messages do not support amino")
 }
 
 // GetSigners implements sdk.Msg
@@ -94,16 +78,6 @@ func NewMsgChannelOpenTry(
 		ProofHeight:         proofHeight,
 		Signer:              signer,
 	}
-}
-
-// Route implements sdk.Msg
-func (msg MsgChannelOpenTry) Route() string {
-	return host.RouterKey
-}
-
-// Type implements sdk.Msg
-func (msg MsgChannelOpenTry) Type() string {
-	return "channel_open_try"
 }
 
 // ValidateBasic implements sdk.Msg
@@ -140,12 +114,6 @@ func (msg MsgChannelOpenTry) ValidateBasic() error {
 	return msg.Channel.ValidateBasic()
 }
 
-// GetSignBytes implements sdk.Msg. The function will panic since it is used
-// for amino transaction verification which IBC does not support.
-func (msg MsgChannelOpenTry) GetSignBytes() []byte {
-	panic("IBC messages do not support amino")
-}
-
 // GetSigners implements sdk.Msg
 func (msg MsgChannelOpenTry) GetSigners() []sdk.AccAddress {
 	signer, err := sdk.AccAddressFromBech32(msg.Signer)
@@ -174,16 +142,6 @@ func NewMsgChannelOpenAck(
 	}
 }
 
-// Route implements sdk.Msg
-func (msg MsgChannelOpenAck) Route() string {
-	return host.RouterKey
-}
-
-// Type implements sdk.Msg
-func (msg MsgChannelOpenAck) Type() string {
-	return "channel_open_ack"
-}
-
 // ValidateBasic implements sdk.Msg
 func (msg MsgChannelOpenAck) ValidateBasic() error {
 	if err := host.PortIdentifierValidator(msg.PortId); err != nil {
@@ -206,12 +164,6 @@ func (msg MsgChannelOpenAck) ValidateBasic() error {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "string could not be parsed as address: %v", err)
 	}
 	return nil
-}
-
-// GetSignBytes implements sdk.Msg. The function will panic since it is used
-// for amino transaction verification which IBC does not support.
-func (msg MsgChannelOpenAck) GetSignBytes() []byte {
-	panic("IBC messages do not support amino")
 }
 
 // GetSigners implements sdk.Msg
@@ -240,16 +192,6 @@ func NewMsgChannelOpenConfirm(
 	}
 }
 
-// Route implements sdk.Msg
-func (msg MsgChannelOpenConfirm) Route() string {
-	return host.RouterKey
-}
-
-// Type implements sdk.Msg
-func (msg MsgChannelOpenConfirm) Type() string {
-	return "channel_open_confirm"
-}
-
 // ValidateBasic implements sdk.Msg
 func (msg MsgChannelOpenConfirm) ValidateBasic() error {
 	if err := host.PortIdentifierValidator(msg.PortId); err != nil {
@@ -269,12 +211,6 @@ func (msg MsgChannelOpenConfirm) ValidateBasic() error {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "string could not be parsed as address: %v", err)
 	}
 	return nil
-}
-
-// GetSignBytes implements sdk.Msg. The function will panic since it is used
-// for amino transaction verification which IBC does not support.
-func (msg MsgChannelOpenConfirm) GetSignBytes() []byte {
-	panic("IBC messages do not support amino")
 }
 
 // GetSigners implements sdk.Msg
@@ -300,16 +236,6 @@ func NewMsgChannelCloseInit(
 	}
 }
 
-// Route implements sdk.Msg
-func (msg MsgChannelCloseInit) Route() string {
-	return host.RouterKey
-}
-
-// Type implements sdk.Msg
-func (msg MsgChannelCloseInit) Type() string {
-	return "channel_close_init"
-}
-
 // ValidateBasic implements sdk.Msg
 func (msg MsgChannelCloseInit) ValidateBasic() error {
 	if err := host.PortIdentifierValidator(msg.PortId); err != nil {
@@ -323,12 +249,6 @@ func (msg MsgChannelCloseInit) ValidateBasic() error {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "string could not be parsed as address: %v", err)
 	}
 	return nil
-}
-
-// GetSignBytes implements sdk.Msg. The function will panic since it is used
-// for amino transaction verification which IBC does not support.
-func (msg MsgChannelCloseInit) GetSignBytes() []byte {
-	panic("IBC messages do not support amino")
 }
 
 // GetSigners implements sdk.Msg
@@ -357,16 +277,6 @@ func NewMsgChannelCloseConfirm(
 	}
 }
 
-// Route implements sdk.Msg
-func (msg MsgChannelCloseConfirm) Route() string {
-	return host.RouterKey
-}
-
-// Type implements sdk.Msg
-func (msg MsgChannelCloseConfirm) Type() string {
-	return "channel_close_confirm"
-}
-
 // ValidateBasic implements sdk.Msg
 func (msg MsgChannelCloseConfirm) ValidateBasic() error {
 	if err := host.PortIdentifierValidator(msg.PortId); err != nil {
@@ -386,12 +296,6 @@ func (msg MsgChannelCloseConfirm) ValidateBasic() error {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "string could not be parsed as address: %v", err)
 	}
 	return nil
-}
-
-// GetSignBytes implements sdk.Msg. The function will panic since it is used
-// for amino transaction verification which IBC does not support.
-func (msg MsgChannelCloseConfirm) GetSignBytes() []byte {
-	panic("IBC messages do not support amino")
 }
 
 // GetSigners implements sdk.Msg
@@ -419,11 +323,6 @@ func NewMsgRecvPacket(
 	}
 }
 
-// Route implements sdk.Msg
-func (msg MsgRecvPacket) Route() string {
-	return host.RouterKey
-}
-
 // ValidateBasic implements sdk.Msg
 func (msg MsgRecvPacket) ValidateBasic() error {
 	if len(msg.ProofCommitment) == 0 {
@@ -437,12 +336,6 @@ func (msg MsgRecvPacket) ValidateBasic() error {
 		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "string could not be parsed as address: %v", err)
 	}
 	return msg.Packet.ValidateBasic()
-}
-
-// GetSignBytes implements sdk.Msg. The function will panic since it is used
-// for amino transaction verification which IBC does not support.
-func (msg MsgRecvPacket) GetSignBytes() []byte {
-	panic("IBC messages do not support amino")
 }
 
 // GetDataSignBytes returns the base64-encoded bytes used for the
@@ -461,11 +354,6 @@ func (msg MsgRecvPacket) GetSigners() []sdk.AccAddress {
 	return []sdk.AccAddress{signer}
 }
 
-// Type implements sdk.Msg
-func (msg MsgRecvPacket) Type() string {
-	return "recv_packet"
-}
-
 var _ sdk.Msg = &MsgTimeout{}
 
 // NewMsgTimeout constructs new MsgTimeout
@@ -481,11 +369,6 @@ func NewMsgTimeout(
 		ProofHeight:      proofHeight,
 		Signer:           signer,
 	}
-}
-
-// Route implements sdk.Msg
-func (msg MsgTimeout) Route() string {
-	return host.RouterKey
 }
 
 // ValidateBasic implements sdk.Msg
@@ -506,12 +389,6 @@ func (msg MsgTimeout) ValidateBasic() error {
 	return msg.Packet.ValidateBasic()
 }
 
-// GetSignBytes implements sdk.Msg. The function will panic since it is used
-// for amino transaction verification which IBC does not support.
-func (msg MsgTimeout) GetSignBytes() []byte {
-	panic("IBC messages do not support amino")
-}
-
 // GetSigners implements sdk.Msg
 func (msg MsgTimeout) GetSigners() []sdk.AccAddress {
 	signer, err := sdk.AccAddressFromBech32(msg.Signer)
@@ -519,11 +396,6 @@ func (msg MsgTimeout) GetSigners() []sdk.AccAddress {
 		panic(err)
 	}
 	return []sdk.AccAddress{signer}
-}
-
-// Type implements sdk.Msg
-func (msg MsgTimeout) Type() string {
-	return "timeout_packet"
 }
 
 // NewMsgTimeoutOnClose constructs new MsgTimeoutOnClose
@@ -541,11 +413,6 @@ func NewMsgTimeoutOnClose(
 		ProofHeight:      proofHeight,
 		Signer:           signer,
 	}
-}
-
-// Route implements sdk.Msg
-func (msg MsgTimeoutOnClose) Route() string {
-	return host.RouterKey
 }
 
 // ValidateBasic implements sdk.Msg
@@ -569,12 +436,6 @@ func (msg MsgTimeoutOnClose) ValidateBasic() error {
 	return msg.Packet.ValidateBasic()
 }
 
-// GetSignBytes implements sdk.Msg. The function will panic since it is used
-// for amino transaction verification which IBC does not support.
-func (msg MsgTimeoutOnClose) GetSignBytes() []byte {
-	panic("IBC messages do not support amino")
-}
-
 // GetSigners implements sdk.Msg
 func (msg MsgTimeoutOnClose) GetSigners() []sdk.AccAddress {
 	signer, err := sdk.AccAddressFromBech32(msg.Signer)
@@ -582,11 +443,6 @@ func (msg MsgTimeoutOnClose) GetSigners() []sdk.AccAddress {
 		panic(err)
 	}
 	return []sdk.AccAddress{signer}
-}
-
-// Type implements sdk.Msg
-func (msg MsgTimeoutOnClose) Type() string {
-	return "timeout_on_close_packet"
 }
 
 var _ sdk.Msg = &MsgAcknowledgement{}
@@ -608,11 +464,6 @@ func NewMsgAcknowledgement(
 	}
 }
 
-// Route implements sdk.Msg
-func (msg MsgAcknowledgement) Route() string {
-	return host.RouterKey
-}
-
 // ValidateBasic implements sdk.Msg
 func (msg MsgAcknowledgement) ValidateBasic() error {
 	if len(msg.ProofAcked) == 0 {
@@ -631,12 +482,6 @@ func (msg MsgAcknowledgement) ValidateBasic() error {
 	return msg.Packet.ValidateBasic()
 }
 
-// GetSignBytes implements sdk.Msg. The function will panic since it is used
-// for amino transaction verification which IBC does not support.
-func (msg MsgAcknowledgement) GetSignBytes() []byte {
-	panic("IBC messages do not support amino")
-}
-
 // GetSigners implements sdk.Msg
 func (msg MsgAcknowledgement) GetSigners() []sdk.AccAddress {
 	signer, err := sdk.AccAddressFromBech32(msg.Signer)
@@ -644,9 +489,4 @@ func (msg MsgAcknowledgement) GetSigners() []sdk.AccAddress {
 		panic(err)
 	}
 	return []sdk.AccAddress{signer}
-}
-
-// Type implements sdk.Msg
-func (msg MsgAcknowledgement) Type() string {
-	return "acknowledge_packet"
 }