+++ sei-ibc-go/modules/core/04-channel/types/msgs_test.go
@@ -315,12 +315,6 @@ func (suite *TypesTestSuite) TestMsgChannelCloseConfirmValidateBasic() {
 	}
 }
 
-func (suite *TypesTestSuite) TestMsgRecvPacketType() {
-	msg := types.NewMsgRecvPacket(packet, suite.proof, height, addr)
-
-	suite.Equal("recv_packet", msg.Type())
-}
-
 func (suite *TypesTestSuite) TestMsgRecvPacketValidateBasic() {
 	testCases := []struct {
 		name    string